requirements.txt
# # Core Machine Learning & Vision
# torch>=2.0.0
# torchvision>=0.15.0

# Image Processing & Augmentation
Pillow>=9.5.0
opencv-python-headless>=4.8.0
albumentations>=1.3.1

# Visualization & Evaluation
matplotlib>=3.7.0

# Utility
tqdm>=4.65.0
awscli>=1.29.0
rasterio